refactor: move PageNotificationLegacyStorageSetup to data-storage and update imports
