+++ apps/web/src/pages/data-marts/list/index.ts
@@ -1,2 +1,1 @@
 export * from './DataMartsPage';
-export * from './PageNotificationLegacyStorageSetup';
